Update GetAllTestInfos func with param issueType
The issueType by default will be ti001 as 'to be investigated', if
provided as empty all failed type will be returned.

Also add the pageSize param which by default is 20.

Update the api query with filter.eq.hasChildren=false to not include the
testsuite parent in failed items.

Add pre-commit config in the project and also do the mass update
follow go lint.

Signed-off-by: Wayne Sun <gsun@redhat.com>